requirements.txt
-r run-requirements.txt
pytest~=4.6.7 # needed for python 2.7+3.4
pytest-cov>=2.8.1
pytest-randomly==1.2.3 # needed for python 2.7+3.4
Flask-Testing==0.8.0
pylint
pylint-flask-sqlalchemy
